+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-## future
+## v0.2.4
 ### Changes
 - Change how memory is measured on linux.
-- Change the time format. It can now use minuts, hours and days.
+- Change the time format. It can now use minutes, hours and days.
 
 ## v0.2.3
 - Update help.